update chuc nang phan quyen, CRUD Customer
